python_ai_pipeline/requirements.txt
# AI Motor Anomaly Detection - Python Dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
tensorflow>=2.13.0
pyserial>=3.5
tqdm>=4.62.0
h5py>=3.1.0
plotly>=5.0.0
jupyter>=1.0.0
